+++ R/VennTable.R
@@ -17,6 +17,7 @@
 #'                  by default: \code{c("1","TRUE","Yes","yes","N/A")}
 #' @param unchecked Vector of character strings that represent unchecked values;
 #'                  by default: \code{c("0","FALSE","No","no","not N/A")}
+#' @param sort      Sort variables by frequency?
 #' 
 #' @details
 #' The column totals ignore missing values.
@@ -43,7 +44,8 @@
 #' @export
 #'
 VennTable <- function(x,markdown=FALSE,NAcode="-",
-  unchecked=c("0","FALSE","No","no","not N/A"),checked=c("1","TRUE","Yes","yes","N/A")) {
+  unchecked=c("0","FALSE","No","no","not N/A"),checked=c("1","TRUE","Yes","yes","N/A"),
+  sort=TRUE) {
   
   TotalSampleSize <- sum(x$n)
   
@@ -61,24 +63,33 @@ VennTable <- function(x,markdown=FALSE,NAcode="-",
     paste0("&nbsp;|",header,"\n",dashes,"\n",body,collapse="\n")
   }
   
-  mat <- as.matrix(x[,-c(1,2)])
+  mat <- as.matrix(x[,-c(1,2),drop=FALSE])
   # Convert logical values to 0's and 1's
   mat[mat %in% unchecked] <- "0"  
   mat[mat %in% checked] <- "1"
   mode(mat) <- "numeric"
   mat <- mat*x$n  # Note that this relies on column recycling
-  count <- apply(mat,2,sum,na.rm=TRUE)
-  xmat <- as.matrix(x)
+
+  countByVar <- apply(mat,2,sum,na.rm=TRUE)
+  if (sort) {
+    o <- rev(order(countByVar))
+    mat <- mat[,o,drop=FALSE]
+    countByVar <- countByVar[o]
+  }     
+  varsorted_mat <- as.matrix(cbind(x[,c(1,2),drop=FALSE],mat))
+  mode(mat) <- "character"
   if (markdown) {
-    xmat[,-(1:2)][is.na(xmat[,-(1:2)])] <- NAcode
-    xmat[,-(1:2)][xmat[,-(1:2)] %in% unchecked] <- ""
-    xmat[,-(1:2)][xmat[,-(1:2)] %in% checked] <- "&#10004;"
+    varsorted_mat[,-(1:2)][is.na(varsorted_mat[,-(1:2)])] <- NAcode
+    varsorted_mat[,-(1:2)][varsorted_mat[,-(1:2)]=="0"] <- ""
+    varsorted_mat[,-(1:2)][
+      varsorted_mat[,-(1:2)]!="" & varsorted_mat[,-(1:2)]!=NAcode] <- "&#10004;"
   }
-  xmat <- rbind(xmat,c(TotalSampleSize,100,rep("",ncol(mat))))
-  xmat <- rbind(xmat,c("","",count))
-  xmat <- rbind(xmat,c("","",round(100*count/sum(x[,1]))))
+  #browser()
+  xmat <- rbind(varsorted_mat,c(TotalSampleSize,100,rep("",ncol(mat))))
+  xmat <- rbind(xmat,c("","",countByVar))
+  xmat <- rbind(xmat,c("","",round(100*countByVar/sum(x[,1]))))
   rownames(xmat) <- c(rep("",nrow(x)),"Total","N","pct")
-  
+
   if (markdown) {
     colnames(xmat)[colnames(xmat)=="pct"] <- "%"
     rownames(xmat)[rownames(xmat)=="pct"] <- "%"